feat:add role and query update

+++ xanadu-system/src/main/java/com/besscroft/xanadu/service/impl/UserServiceImpl.java
@@ -5,6 +5,7 @@ import cn.dev33.satoken.stp.SaTokenInfo;
 import cn.hutool.core.util.IdUtil;
 import cn.hutool.core.util.StrUtil;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.besscroft.xanadu.common.constant.RoleConstants;
 import com.besscroft.xanadu.common.constant.SystemConstants;
 import com.besscroft.xanadu.common.converter.UserConverterMapper;
 import com.besscroft.xanadu.common.entity.User;
@@ -57,14 +58,19 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
     }
 
     @Override
-    public List<User> userPage(Integer pageNum, Integer pageSize) {
+    public List<User> userPage(Integer pageNum, Integer pageSize, String role) {
         PageHelper.startPage(pageNum, pageSize);
-        return this.baseMapper.selectPage();
+        return this.baseMapper.selectPage(role);
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void deleteUser(Long userId) {
+        // TODO 适配后需改成上下文中获取，而不是直接查库
+        User user = this.baseMapper.selectById(userId);
+        Assert.notNull(user, "用户不存在！");
+        if (Objects.equals(user.getRole(), RoleConstants.PLATFORM_SUPER_ADMIN))
+            throw new XanaduException("超级管理员不允许被删除！");
         Assert.isTrue(this.baseMapper.deleteById(userId) > 0, "用户删除失败！");
     }
 
@@ -77,6 +83,8 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
     @Transactional(rollbackFor = Exception.class)
     public void addUser(UserAddParam param) {
         User user = UserConverterMapper.INSTANCE.AddParamToUser(param);
+        if (Objects.equals(user.getRole(), RoleConstants.PLATFORM_SUPER_ADMIN))
+            throw new XanaduException("违反规则！超级管理员角色不允许被添加！");
         user.setStatus(SystemConstants.STATUS_NO);
         user.setPassword(SaSecureUtil.sha256(param.getPassword().trim()));
         Assert.isTrue(this.baseMapper.insert(user) > 0, "新增用户失败！");
@@ -86,6 +94,10 @@ public class UserServiceImpl extends ServiceImpl<UserMapper, User> implements Us
     @Transactional(rollbackFor = Exception.class)
     public void updateUser(UserUpdateParam param) {
         User user = UserConverterMapper.INSTANCE.UpdateParamToUser(param);
+        User oldUser = this.baseMapper.selectById(user.getId());
+        if ((!Objects.equals(oldUser.getRole(), RoleConstants.PLATFORM_SUPER_ADMIN) && Objects.equals(user.getRole(), RoleConstants.PLATFORM_SUPER_ADMIN))
+                || (Objects.equals(oldUser.getRole(), RoleConstants.PLATFORM_SUPER_ADMIN) && !Objects.equals(user.getRole(), RoleConstants.PLATFORM_SUPER_ADMIN)))
+            throw new XanaduException("违反规则！更新用户失败！");
         if (StrUtil.isNotBlank(param.getPassword())) {
             user.setPassword(SaSecureUtil.sha256(param.getPassword().trim()));
         }
